+++ solutions/copick/project_from_dataportal/CHANGELOG.md
@@ -4,6 +4,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.9] - 2024-06-11
+Fix config file adjustment
+
+## [0.1.8] - 2024-06-11
+Fix typo in s3 path
+
+## [0.1.7] - 2024-06-11
+Save into overlay then update config
+
+## [0.1.6] - 2024-06-11
+Fix zarr copying
+
+## [0.1.5] - 2024-06-11
+Fix zarr copying
+
+## [0.1.4] - 2024-06-11
+Fix zarr module
+
+## [0.1.3] - 2024-06-11
+Add missing matplotlib dep
+
+## [0.1.2] - 2024-06-11
+Update s3 usage
+
+## [0.1.10] - 2024-06-11
+Fix copick paths
+
+## [0.1.1] - 2024-06-11
+Update colors and save directly into zarr store
+
+## [0.1.0] - 2024-06-11
+Deploy revised copick-oriented version
+
 ## [0.0.8] - 2024-04-18
 Upgrade to latest pixel embedding generator